Add files via upload
Inside the zip folder, there are two folders: one containing the two-player game project files, and the other containing the weekly tasks. The two-player game project includes all the code files required to run the programme, a README pdf file describing the idea and development stages of the game, and all additional assets needed to run the game (font and image files used).